Remove pragmas on warning levels
It's not a library's job to specify at what level warnings should be shown,
as this interferes with the downstream project's decisions on this
matter.

It can of course still use pragmas to locally disable warnings in
headers (by pushing and popping the warning levels) or in C++ files.

Instead, we add a global flag in CMake to set the warning level, as this
is only valid while building Lass, and not when using it in a downstream
project.

We set the warning level to the highest level /W4, and will fix or
silence warnings accordingly.