Remove pragmas on warning levels
It's not a library's job to specify at what level warnings should be shown,
as this interferes with the downstream project's decisions on this
matter.

It can of course still use pragmas to locally disable warnings in
headers (by pushing and popping the warning levels) or in C++ files.

Instead, we'll add a global flag in CMake to set the warning level at
level 3, as this is only valid while building Lass, and not when using
it in a downstream project.

Later, we will increase warning level to 4, and fix or silence the extra
warnings during the Lass build.