Replace usage of LASS_UNUSED by [[maybe_unused]]
On MSVC, LASS_UNUSED(identifier) compiles to just identifier. It's only
on GCC and Clang that it added the __unused__ attribute. Now at warning
level /W4, also MSVC complains:

  warning C4100: 'identifier': unreferenced formal parameter
  warning C4189: 'identifier': local variable is initialized but not referenced

We can fix this with [[maybe_unused]], but not as part of the
LASS_UNUSED macro, as [[maybe_unused]] must be placed before the type,
not before or after the identifier.

Curiously enouogh, we don't need to replace all of them, so it seems not
all "unused" variables are really unused...  To be investigated.