Add CTest output on failure
When running the test suite through CMake (by building RUN_TESTS are
make test), you don't get to see the output of the tests. You just see
something like this:

       Start  1: test_io_arg_parser
  1/82 Test  #1: test_io_arg_parser ................. Passed  0.00 sec
       Start  2: test_io_binary_stream
  2/82 Test  #2: test_io_binary_stream ........... ***Failed  0.00 sec
       Start  3: test_io_file_attributes
  3/82 Test  #3: test_io_file_attributes ............ Passed  0.01 sec

This makes it hard to debug failing unittests on CI builds, where it's
hard to get to the output log files, or to rerun unittests manually.

This flag will print the output of failing tests, for easier debugging.