Use Woop ray-triangle intersection in TriangleMesh3D
The ray-triangle intersection from [1] is designed to be watertight
which makes it very suitable to be used in a triangle mesh, to ensure a
robust ray-mesh intersection in the presence of small or needle-like
triangles. This algorithm has also been adopted in PBRT 3ed [2]

We only implemented the ray-triangle intersection. The paper also how
the bounds need to be enlarged during BVH traversal to conservatively
bound the triangles visited, but this has not yet been implemented,
neither does it seem to have been done in [2]

[1] Sven Woop, Carsten Benthin, and Ingo Wald. "Watertight Ray/Triangle
    Intersection." Journal of Computer Graphics Techniques (JCGT),
    2(1):65--82, 2019. https://jcgt.org/published/0002/01/05/
[2] M. Phar, W. Jakob, G. Humphreys, Physically Based Rendering: From
    Theory To Implementation, 3rd ed., San Francisco, CA, USA:
    Morgan Kaufmann, 2017. ISBN 978-0-12-800645-0
    https://www.pbr-book.org/3ed-2018/Shapes/Triangle_Meshes#TriangleIntersection