Hide Python dispatcher functions from interface
On Linux, hide the symbols of the automatically generated C functions
that are used as dispatchers to hook exported C++ functions to a Python
type, using __attribute__ ((visibility ("hidden"))) [1]. On Linux, these
symbols are public by default (not on Windows), and as these C functions
don't have namespaces, a nameclash can happen in some circumstances,
with weird runtime errors as consequence, as demonstrated in following
example:

In liar [2], there are two transformation classes:
liar.scenery.Transformation for SceneObject objects, and
a SceneObject with the former, you get the following error:

  >>> import liar
  >>> tr = liar.Transformation3D.identity()
  >>> sky = liar.scenery.LightSky()
  >>> liar.scenery.Transformation(sky, tr)
  Traceback (most recent call last):
    File "<stdin>", line 1, in <module>
  TypeError: Bad Argument on 1th position: scenery.LightSky not convertable to kernel.Medium

LightSky is indeed not a Medium, because it's a SceneObject subclass.
Let's try with a medium instead:

  >>> fog = liar.mediums.Fog()
  >>> liar.scenery.Transformation(fog, tr)
  <scenery.Transformation object at 0x558ad21fc898>

This "works", but let's try to access an attribute:

  >>> _.child
  Traceback (most recent call last):
    File "<stdin>", line 1, in <module>
  TypeError: scenery.Transformation not convertable to mediums.Transformation

What happened is that both types have automatically generated
'extern "C"' functions on the "same" line numbers, but in different
files, in different modules even. Because the automatically determined
function names don't include the type's namespace and they're not in a
namespace themselves (they're C functions), and only use the linenumbers
to make them unique, we end up with exactly the same function names for
scenery.Transformation and mediums.Transformation, for example:

  lassPyImpl_constructor_Transformation33

If both types would have been exported in the same module, the linker
would have complained having more than one defintion for these symbols,
but as they are in different modules, the conflict only happens at
runtime. Only one set "survives" and in this case the functions of
mediums.Transformation are used for both types.

However, these functions don't need to be visible in the global symbol
table, they are private to each module. On GCC and Clang, we define a
macro LASS_DLL_LOCAL to __attribute__ ((visibility ("hidden"))) [1] and
use that to hide these symbols. On Windows, they are already hidden by
default, so LASS_DLL_LOCAL is empty.

I also considered using "-fvisibility=hidden" instead, and have
LASS_DLL_EXPORT and LASS_DLL_IMPORT be defined to
__attribute__ ((visibility ("default"))) [1], but that doesn't help for
modules that use these export macros that don't use
"-fvisibility=hidden". Explictly marking them as hidden is better in
this case.

Defining LASS_DLL_EXPORT to __attribute__ ((visibility ("default")))
also gave some troubles when combining it with the [[deprecated]]
attribute, for example in random.h, and the internet is not really
helpful in find a solution here [3]:

  lass/lass/num/random.h:100:9: error: expected identifier before ‘[’ token
    100 |         [[deprecated("RandomParkMiller is deprecated, use std::minstd_rand0 instead, or upgrade to std::minstd_rand")]]
        |         ^
  lass/lass/num/random.h:100:9: warning: attribute ignored [-Wattributes]
  lass/lass/num/random.h:100:9: note: an attribute that appertains to a type-specifier is ignored

Defining using the C++11 attribute syntax to define LASS_DLL_EXPORT and
LASS_DLL_IMPORT to [[gnu::visibility("default")]] instead, that gives
different problems, for example in python/exception.h (LASS_PYTHON_DLL
expands to [[gnu::visibility("default")]]), so I left it alone:

  lass/lass/python/exception.h:66:16: error: standard attributes in middle of decl-specifiers
     66 |         static LASS_PYTHON_DLL const std::string LASS_CALL extractMessage(PyObject* type, PyObject* value = 0);

[1] https://gcc.gnu.org/wiki/Visibility
[2] https://liar.bramz.net/
[3] https://stackoverflow.com/a/73435383
[4] https://gcc.gnu.org/onlinedocs/gcc/Attribute-Syntax.html#Attribute-Syntax