Add optional filter to TriangleMesh intersect
This allows to implement an alpha-mask more efficiently in ray tracing
applications. An alpha-mask texture can be used to add finer countours
to a surface mesh while keeping coarser triangles.

While traversing the spatial tree in TriangleMesh, if a candidate
triangle intersection is found, the filter is being evaluated. Only if
it returns true, the intersection is retained and the intersection with
the smallest t > tMin is returned. If it returns false, the triangle is
ignored and a more distant intersection will be returned.

Without the filtered intersection method, the alpha mask must be
implemented outside TriangleMesh, on the triangle intersection closest
to tMin. If the alpha mask returns false for that intersection, you must
restart the intersect method with updated tMin (equal to t of the
previously returned but rejected triangle), until you find a triangle of
which the alpha mask returns true, or until no more intersections can be
found. Each you, traversal of the tree must restart from the root. With
this new method, you only need one traversal of the tree, as the alpha
mask can be evaluated while traversing it.