Relax PyExportTraits for floating points
Before, PyExportTraits<float> would not support (minus) infinity with
following errors:

  TypeError: not a float: overflow: inf > 3.40282e+38
  TypeError: not a float: underflow: -inf < -3.40282e+38

In general, it would not accept numbers larger than its max value:

  TypeError: not a float: overflow: 1e+40 > 3.40282e+38

In C++, a static_cast would map 1e+40 to +inf. This is a lossy
round-trip, but so is casting 1.0000000001 to 1.f or 1e-50 to 0.f,
or 1e-10 to something not exactly 1e-10 (see unittest).

So I've relaxed this to not take into account min and max float value,
and simply perform a static_cast from double to float.

In same fashion, I've also relaxed the way integers are casted, in
accordance with Python's ASSIGN_DOUBLE implementation in
mathmodule.c [1].

And so, we also support custom float objects with __float__

By defining LASS_USE_OLD_EXPORTRAITS_FLOAT=1 (in CMake), you can revert
back to the old behavior.

[1] https://github.com/python/cpython/blob/v3.12.4/Modules/mathmodule.c#L211-L231