Make test_suite run with static Python library
On Linux, when linking against Python as a static library,
e.g. libpython3.12.a, then you will get following error when running the
test_suite:

  Traceback (most recent call last):
    File "/home/bramz/coding/Python-3.12.4/dist/lib/python3.12/datetime.py", line 2, in <module>
      from _datetime import *
  ImportError: /home/bramz/coding/Python-3.12.4/dist/lib/python3.12/lib-dynload/_datetime.cpython-312-x86_64-linux-gnu.so: undefined symbol: PyFloat_Type

Extensions don't link against the Python library, and are getting their
symbols from the interpreter they're loaded in. However, when the
executable test_suite links against Python as a static library, those
symbols will not be exposed to the extensions loaded with dlopen. To fix
this, you need to add the -export-dynamic linker option [1]

[1] https://stackoverflow.com/a/67894308