Use isolated python for test_suite
Isolate the embedded Python in the test_suite from the environment to
mitigate interference from other Python environments in the system.
In this Isolated Configuration [1], it is recommended to explicitly set
PyConfig.home, which we set to the base_prefix obtained from the used
Python. We query sys.base_prefix instead of sys.prefix so that in case
of a virtual environment, we obtain the prefix with the standard
library.

[1] https://docs.python.org/3.12/c-api/init_config.html#init-isolated-conf