Export tracked sources only
git.included_files() also lists untracked files, which we don't really
want to export as sources in the Conan package. We only need the files
that are also tracked in the Git repository.

git.included_files() runs following command to list files, which are
_all_ files not ignored by .gitignore [1]:

  git ls-files --full-name --others --cached --exclude-standard

If we replace this by the output of simply `git ls-files`, we get the
files that we want [2]:

- --full-name is not required, as we run this from the root of the repo,
  or at least relative to the conanfile.py
- --others is what yields the untracked files, so this we want to omit
- --cached is the default if we don't specify any other of following
  options: -c/-s/-d/-o/-u/-k/-m/--resolve-undo
  It also includes staged files, in case we want to make a "dirty"
  export with a previously-untracked file: you can stage the file
  without commiting it.
- --exclude-standard was only needed for the --others flag to not return
  files being ignored by .gitignore.

[1] https://docs.conan.io/2/reference/tools/scm/git.html#conan.tools.scm.git.Git.included_files
[2] https://git-scm.com/docs/git-ls-files