Fix default target name for CMakeDeps
Fix the default CMake target name that Conan will use to declare a
dependency on Lass in a dependent's CMakeDeps-generated CMake config.

To illustrate the problem:

  Take two Conan packages foo and bar. foo depends on Lass and bar
  depends on foo:

    Lass <-- foo <-- bar

  (bar could also be foo's test_package)

  foo simply requires "lass" (in Conan 2, the package name is lowercase),
  uses CMakeToolchain and CMakeDeps, and will use find_package(Lass) to
  use Lass in it's CMakeLists.txt. So far, no problem, foo will build as
  normal.

  Say also that foo does not generate its own foo-config.cmake, but uses
  package_info so that Conan will be able to generate this.

  bar simply requires "foo", also uses CMakeToolchain and CMakeDeps. In
  its build's generators folder, there will be a foo-config.cmake file
  generated by Conan. In foo-Target-<build_type>.cmake, it will declare
  an imported target foo, with a CMake depencency on "lass::lass". This
  CMake target does not exist, and bar will fail to build:

    The link interface of target "foo_DEPS_TARGET" contains:
      lass::lass
    but the target was not found.  Possible reasons include:
      * There is a typo in the target name.
      * A find_package call is missing for an IMPORTED target.
      * An ALIAS target is missing.

The reason why this fails is that while generating foo-config.cmake,
CMakeDeps does not know that the CMake targets of Lass are Lass::lass
and Lass::lass_python, using a capital L for the namespace, because
those targets are not generated by Conan CMakeDeps, but rather by Lass'
CMakeLists.txt instead.  Instead, not being told otherwise, it simply
assumes a target "<package>::<package>" that links to all libs of the
package: both lass _and_ lass_python. And in Conan 2, the package is
lowercase "lass".

We can fix that by simply setting the cmake_target_name property in
package_info:

  self.cpp_info.set_property("cmake_target_name", "Lass::lass")

In bar's generators folder, this will now result in a CMakeDeps
generated import target for foo that will have a proper CMake dependency
on Lass::lass. If we don't need Python, this will build.

However, if foo depends on Lass::lass_python instead, we will have no
way to express this. We could try to declare the following in foo's
package_info:

  self.cpp_info.requires = ["Lass::lass_python"]

However, we'll get following error (with Conan 2):

  ERROR: pkg/1.0: required component package 'Lass::' not in dependencies

That's because the dependencies in package_info's requires are Conan
components, not CMake targets. They use the same :: separator, but
they're "<package>::<component>" instead of "<namespace>::<target>", and
the Conan 2 package name of Lass is lowercase "lass".

But even with the lowercase "lass::lass_python", we still get the error:

  ERROR: Error in generator 'CMakeDeps': error generating context for 'foo/1.0': Component 'lass::lass_python' not found in 'lass' package requirement

So we also need to declare two components "lass" and "lass_python" in
package_info.

And because we can't mix component with normal non-component
cpp_info declarations, we need to need to duplicate the layout per
component as well. Except for the cmake_find_mode property, which then
should _not_ be defined on the component.

Lastly, we still set the cmake_target_name property to "Lass::lass", so
that by default, packages that depend on Lass will _only_ depend on
Lass::lass.

If your package depends on lass_python, and you let CMakeDeps generate
its config file, you'll have to add "lass::lass_python" to its requires
in package_info:

  def package_info(self):
    self.cpp_info.requires = ["lass::lass_python"]