Keep global cpp info dirs
If we only define the directories on the components, then on the package
level, they will still have values. These defaults are fine for
includedirs, libdirs and bindirs in package mode; but they'll be wrong
in editable mode. And also builddirs will be wrong, so LassConfig.cmake
will not be found.