Keep global cpp info dirs
If we only define the directories on the components, then on the package
level, they will still have their default values. These defaults are
fine for includedirs, libdirs and bindirs in package mode; but they'll
all be wrong in editable mode. And also builddirs will be wrong, even in
normal package mode, so LassConfig.cmake will not be found.