Improve precision of Ray3D/Sphere3D intersection
In Debug mode on Windows, the unit test testSpatObjectTrees<float,3>
in test_spat_object_trees.cpp was failing because of an assertion error
in AabpTree:

  lass/spat\aabp_tree.inl(681): ASSERT FAILURE: 'tCandidate > tMin &&
  tCandidate >= tNear * (1 - 1e-6f) && tCandidate <= tFar * (1 + 1e-6f)'

with:

  tCandidate = 2511.18359
  tMin = 0.00000000
  tNear = 2511.26074
  tFar = 2525.92407

This was odd. tNear and tFar bound the volume that contains all objects
in the node, and thus tCandidate should always be between these values,
give or take a tolerance.

Using Higham's gamma notation [4], it was found that the rounding-error
on the floating point operations for tNear and tFar is γ_3, and γ_8 for
the ray-sphere intersection, both are smaller than the 1e-6f being used
here, even for single precision floating point. So the tolerance is just
fine.

However, despite using double precision floating points internally in
the ray-sphere intersection algorithm, it was found that the
computation of the discriminant observed heavy cancelation, due to a
small radius compared to cs = the distance of the ray's support point to
the sphere's center. I.e. only a few bits of the radius were represented
in c, and sqr(b) and c were both large and comparable, resulting in
heavy cancelation:

  cs = ray.support() - sphere.center();
  b = dot(cs, ray.direction());
  c = cs.squaredNorm() - sqr(sphere.radius());
  discriminant = sqr(b) - c

I've found following paper [1] describing two little-known
reformulations to improve the precision of the result. The most
important for this case is the geometry based interpretation of the
discriminant [2],

  l = cs - b * d;
  discriminant = sqr(r) - l.squaredNorm();

Even if l is still small compared to cs and b * d, the difference is
taken before the squaring, preserving more bits. And sqr(r) also
preserves more bits in the final discriminant, as we always have
sqr(r) > l.squaredNorm() _if_ there's an intersection.

The other improvement is less relevant for this particular case, but it
improves computation of the final t1 and t2 in the other extreme case
when the ray's support point is close to a very large sphere, and b is of
comparable size to sqrt(discriminant). It avoids the cancelation
depending on the sign of b [3, 4].

Still in the case of the ray's support point close to a very large
sphere, we have cancelation in computation of c. It was found that with
single-precision floats, we were still failing the testcase, and the
computation of c turned out to be the crucial part, which is why we're
doing this part in double precision.

The exact case on which the AabpTree's assertion was failing has been
reproduced in test_prim_ray_3d_sphere_3d.cpp. Using a multi-precision
math library [5], the expected values for t1 and t2 has been calculated
exactly in a python script test_suite\test_prim_ray_3d_sphere_3d.py

[1] Haines, E., Günther, J., Akenine-Möller, T. (2019).
    Precision Improvements for Ray/Sphere Intersection.
    In: Haines, E., Akenine-Möller, T. (eds) Ray Tracing Gems.
    Apress, Berkeley, CA. https://doi.org/10.1007/978-1-4842-4427-2_7
[2] Hearn, D. D., and Baker, M. P. Computer Graphics with OpenGL,
    3rd ed. Pearson, 2004
[3] Press, W. H., Teukolsky, S. A., Vetterling, W. T., and Flannery, B. P.
    Numerical Recipes: The Art of Scientific Computing, 3rd ed.
    Cambridge University Press, 2007.
[4] N. J. Higham, Accuracy and Stability of Numerical Algorithms,
    2nd ed., Philadelphia, PA, USA: Society for Industrial and Applied
    Mathematics, 2002. ISBN 0-89871-521-0
    https://doi.org/10.1137/1.9780898718027
[5] https://mpmath.org/