Parse parameter names of constructors
When parsing a constructor dispatcher, we cannot determine the names of
the constructor parameters, as we never have a function pointer to the
constructor itself, unlike method or function dispatchers. We only know
the argument types, which always need to be explicitly spelled out.

However, when parsing a class declaration, we can scan the whole class
for constructors that do have parameter names, and store that
information for later. Then, when a constructor is exported, we use its
signature to scan for a matching constructor to retrieve the parameter
names.

Matters get more complicated when the exported class is in fact an
instantiated class template, being exported via a shadow class.
The instantiated class "body" will be empty, and we need to find the
class template itself. But the constructor parameter types will likely
also depend on template parameters. We cannot always substitute them for
real parameter types (e.g. in case of pylass::Polynomial), and we'll
be unable to match the signatures and get the real parameter names.