Require Python >= 3.10 and 64-bit for lass_stubgen
Python 3.9 allowed to use built-in types like `list` and `dict` as
generics [1], and Python 3.10 introduced | as the union operator [2].
This makes it cumbersome to support anything before Python 3.10.
At least `Self` [3] we can still get from typing-extensions in 3.10.
And for `StrEnum`, there's also a simple workaround.

As Python 3.9 goes end-of-life in October [4], it's an easy
decision to not support it for lass_stubgen.

Also, libclang module is not supported on 32-bit x86.

[1] https://docs.python.org/3/whatsnew/3.9.html#type-hinting-generics-in-standard-collections
[2] https://docs.python.org/3/whatsnew/3.10.html#new-features-related-to-type-hints
[3] https://docs.python.org/3/whatsnew/3.11.html#pep-673-self-type
[4] https://devguide.python.org/versions/