Support precompiled headers in lass_stubgen
This especially speeds up stub generation on machines with fewer cores.

Don't bother scanning the AST of the PCH, as we we'll be scanning the
same parts for each individual C++ file anyway. I haven't found a way
for each C++ file to skip the first part of the AST (that should be the
same as for the PCH) and immediately jump to the unique part. At least,
we can avoid wasting scanning the PCH itself.