Fix refcount of enum when added to module
PyModule_AddObject steals a reference of the object being added [1],
which was not counted for by the enum definition. Then, at exit, in the
destructor of EnumDefinitionBase, it tries to clean up TPyObjPtr type_,
but the refcount will already have dropped to zero, meaning the object
might already have been destroyed and the pointer should no longer be
dereferenced.

The docs recommend replacing it by PyModule_Add or PyModule_AddObjectRef
PyModule_AddObjectRef was choosen because it doesn't steal any
reference, and it was introduced earlier in 3.10, rather than 3.13.

[1] https://docs.python.org/3/c-api/module.html#c.PyModule_AddObject