Pass -rdynamic as link option instead
We added -rdynamic to be able to link to a static build of the Python
library [1], to make sure its symbols are exposed to the extension
modules being loaded by the embedded Python.

But Clang was compiling with:

  clang++: warning: argument unused during compilation: '-rdynamic' [-Wunused-command-line-argument]

And neither GCC nor Clang actually did anything with it. I.e. I still
got the undefined symbols when I was linking to a static Python library.

Well, d'uh! Turns out, we need to pass this as *link* option instead of
*compile* option! :facepalm:

[1] commit 5629f613671fce7aa30698276d659adde1f4d5d7