Add each stubgen file seperately to build system
With the old system, we could not easily control how many files would be
parsed in parallel in a larger build system, while simultaneously
compiler other source files. Especially on lower end devices this could
be really taxing on memory. By adding each stubgen file by its own to
the build system, the build system itself can better control the
parallelization. This is especially evident with Makefiles. MSBuild has
less control over this, but it's at least not worse than how it
parallelizes compilation of source files.

The stubgen tool generates *.d files listing the included files, which
are being used as DEPFILE argument to the custom commands. This will
make sure the stubgen files are rebuild when any of its dependencies
changes. This puts a requirement on CMake 3.21 or newer, for this to
work for Makefiles and Visual Studio projects.

For parallel custom build steps to work with MSBuild, you at least need
to use CMake 3.27. MSBuild will still fallback to sequential execution
if all-but-one build steps of a custom target depend on a single custom
rule. In particular, they all depend on the precompiled header step. To
avoid this, we had to move the precompiled header to a separate target.

We dropped support for the --obj-file and --cache-dir options, as that
that mechanism is now obsolete. Multi-threading is now obsolete as well,
but I've kept that for the time being.