Clean up test_python_embedding.py, and test with mypy and ruff
Make sure all tests are put in a TestCase, so there's no need for manual
checking of stdout.

Make sure mypy and ruff checks are error free.

We moved around with the iter functions, as mypy was rightfully
complaining that the second iter overload on ClassMap would never be
used.