Add --debug-connect to lass_stubgen
This will connect to a listening DAP [0] client for debugging purposes.
It uses debugpy to connect [1] to a listening debugger on the given port.
First start the debugger in listening mode, then build the stubs with
this option.

To start a debugger in listening mode in VSCode, you can use the
following configuration in .vscode/launch.json:

{
  "version": "0.2.0",
  "configurations": [
    {
      "name": "Python: Listen 5678",
      "type": "debugpy",
      "request": "attach",
      "listen": {
        "host": "localhost",
	"port": 5678
      }
    }
  ]
}