Propagate exceptions from NodeVisitor to Parser
If an exception is raised in one of the handlers, and we don't handle
the exception, the following error is printed and parsing continues:

  Exception ignored on calling ctypes callback function: <bound method NodeVisitor.visit_node of <lass_stubgen.parser.NodeVisitor object at 0x0000025F75C3EE60>>

If we capture the exception instead, break out of the visitor recursion,
we can reraise the exception outside the ctypes callback and get proper
exception handling.

Add notes with location of node that was being parsed.