Fix missing VAR_DECL of ModuleDefinition
In some cases, the initializer expression of ModuleDefinition (the
constructor call) does not have its semantic_parent set. Sometimes it's
None, especially for top-level declarations—according to Copilot.

We need this semantic_parent because it's the variable declaration
(VAR_DECL) of the ModuleDefinition that will give us variable name
(cpp_name) by which the module will be referred to by other macro calls.

It's suggested to check the lexical_parent, but that might also be None.

Another suggestion to robustly find the VAR_DECL, as a fallback, is to
scan over the children of the translation_unit, and find the top-level
VAR_DECL node that has the same location as the initializer expression.
This seems to be working happily.