Fix lass_stubgen support for PyExportTraitsEnum
The old-fashioned way to export enums is to derive its PyExportTraits
from PyExportTraitsEnum, which translates the enum to and from a Python
int.

Using _find_first_child instead of _get_first_child, because the
first child node does not need to be the CXX_BASE_SPECIFIER node we're
looking for.