Fix sorting of partially matched export traits
It's not the number of the template parameters that matters, but which
one is more specialized. From cppreference.com, this is explained as [1]

  > Informally "A is more specialized than B" means "A accepts a subset
  > of the types that B accepts".

So we defined a partial ordering on ExportTraits, where `self < other`
means that self is more specialized than other. If they cannot be
compared (because they are for different types, for example) then False
is returned for both `self < other` and `other < self`.

[1] https://cppreference.com/w/cpp/language/partial_specialization.html