Find libclang library for lass_stubgen
Since upgrade to Visual Studio 2022 17.14.8, lass_stubgen started to
fail with following error:

  C:\Program Files\Microsoft Visual Studio\2022\Community\VC\Tools\MSVC\14.44.35207\include\yvals_core.h(908,1): error GC074E59A: static assertion failed: error STL1000: Unexpected compiler version, expected Clang 19.0.0 or newer.

The libclang package [1] that we're using—which has the required
libclang(.dll|.so) embedded—seems to be unmaintained. It has no newer
version than 18.1.

There's the clang package [2] that seems maintained and supports the
newest Clang versions, but that doesn't contain the libclang library.
To use it, we need to find the libclang library elsewhere.

There is a PR [3] that wants to add clang to the Conan center, but
that's still open. On the CMake level, we have no other choice that to
search for it on the build system.

At least on Windows, we can search for libclang.dll in the Visual Studio
installation. This requires that the "C++ Clang Compiler for Windows"
(Microsoft.VisualStudio.Component.VC.Llvm.Clang) component was selected
during installation.

On Linux, if we're building using a Clang compiler, we can use the same
one. Otherwise, we'll assume `clang` is available on `$PATH`. In both
cases, we query the executable to find its resource dir and continue
from there ...

[1] https://pypi.org/project/libclang/
[2] https://pypi.org/project/clang/
[3] https://github.com/conan-io/conan-center-index/pull/24315