Improve matching raw pointers in export traits in lass_stubgen
Without this change, a C++ type like `python::NoNone<RawType*>` would
not properly be matched during stub generation and result in a signature
like this:

  def testNoNoneRaw(bar: RawType | None, returnNone: bool, /) -> RawType | None: ...

Although at runtime, `python::NoNone<RawType*>` would be have as
expected, and raise a `TypeError` when None or nullptr was passed.

The bug originates from the way we try to match `python::NoNone<RawType*>`
to a ExportTraits in generator.py. Though an ExportTraits specialization for
`NoNone<T*>` exists, it would never be matched and the only match would
be the general specialization:

```
   1: RawType | None <- lass::python::NoNone<T> with T=lass::test::RawType *
```

With this fix, we get two possible matches with the second one being
more specific:

```
   1: RawType | None <- lass::python::NoNone<T> with T=lass::test::RawType *
*  2: RawType <- lass::python::NoNone<T *> with T=lass::test::RawType
```

And we finally get the signature that we needed:

  def testNoNoneRaw(bar: RawType, returnNone: bool, /) -> RawType: ...