Streamline dependencies of stubdata custom commands
Before, upstream targets were added as dependencies of each stubdata
file. That causes all source files to be reparsed for stubs even if just
a single file of the upstream target changed. Now, we only add the
upstream targets as dependencies to the custom target _bundling_ the
custom commands for the stubdata files: if just one file changes, then
only the assiocated stubdata file will be exctracted again.

The upstream targets are still dependencies of the custom targets, so
that the invidual custom commands of the custom target are only executed
after completing the upstream targets. This is especially relevant if
the upstream targets (or their dependencies) generate any source files
that are needed during the stub extraction.