Prepend Clang system include dir to lass_stubgen
Now we're relying on an external libclang library, we can also try to
find the associated system include directory which we need if our code
relies on intrinsics. Those headers are not compatible between Clang and
GCC or MSVC, and using the GCC/MSVC headers will lead to "compilation"
errors during the stubdata extraction.

The libclang Python package didn't include them, so we didn't bother
before, and in case of LiAR, I relied on simply using Clang to build
LiAR as well. But now, we can try to find them and prepend them to the
system include dirs ${CMAKE_CXX_IMPLICIT_INCLUDE_DIRECTORIES}.

On Windows, CMAKE_CXX_IMPLICIT_INCLUDE_DIRECTORIES is empty, but the
`INCLUDE` environment variable will be set during lass_stubgen
execution.