Hardcode PyObject* parameters as Any
Raw PyObject* pointers can be used as parameters in exported C++
functions to accept Python objects of some undisclosed type. The
function gets a borrowed reference to the object as agument. I.e. it
doesn't need to—and should not!—decref the pointer, and so there's no
risk of leaking the reference. If it wants to steal a reference, it
needs to incref it first. So this is quite safe, there's no risk of
leaking a reference. PyObject* as return types are not allowed though,
as it's less obvious what needs to happen with the refcount.

Without this fix, PyObject* parameters are typed as `_object | None`.
`_object` is the actual C-type that `PyObject` is a typedef for:

  def testRawPyObject(obj: _object | None, /) -> Any: ...

This fix replaces it by `Any`, just like for `TPyObjPtr`, signaling that
we don't really know what types are accepted as arguments, or what type
is returned (in case of TPyObjPtr).

Why not type-hint it as `object`?  After all, `object` *is* the _exact_
Python type of `PyObject*` in the first place!

That was my first choice too, and it would also have worked too: mypy
would happily allow to pass objects of any type without any complaint.
For the type-checker, all types are subclasses of both `Any` and
`object`. But there's a subtle difference [1,2]:

- If we use `object`, we say that any object of any type is a valid
  argument: we're never going to get a TypeError because we passed an
  object of the wrong type. This is generally true for comparison
  operators, for example, which return `NotImplemented` if two types
  can't be compared, which eventually evaluates to `False`. [3]

- If we use `Any`, we're saying that we don't _know_ what types are
  acceptable or not, and that the type-checker shouldn't bother checking
  it. We might get a TypeError at runtime, if we pass an object of the
  wrong type, but the type-checker will never flag it.

In general, we don't know what types are accepted by PyObject*
arguments, so `Any` seems the better choice.

[1] https://mypy.readthedocs.io/en/latest/dynamic_typing.html#any-vs-object
[2] https://stackoverflow.com/a/39817126
[3] https://stackoverflow.com/a/37557540