Replace the current compiler's private headers with Clang's
Instead of simply prepending libclang's private headers include dir to
the ${CMAKE_CXX_IMPLICIT_INCLUDE_DIRECTORIES}, look for the current
compiler's private headers dir (GCC or Clang), and replace that
directory, in that position, by libclang's.

Simply prepending libclang's include dir works on Windows (where it is
effectively prepended to MSVC's include dirs in the INCLUDE environment
variable when you compile your code with clang-cl), but on Linux it may
cause issues.

For example, when using the PATH_MAX constant, you may encounter
following error during the stub extraction:

    error: use of undeclared identifier 'PATH_MAX'

This seems to caused by Clang's limits.h header being found first.

On one system (Ubuntu 2024.04), here's what the list of default include
directories look like for GCC, using the command
`echo | gcc -xc++ -E -v -`:

    /usr/include/c++/13
    /usr/include/x86_64-linux-gnu/c++/13
    /usr/include/c++/13/backward
    /usr/lib/gcc/x86_64-linux-gnu/13/include
    /usr/local/include
    /usr/include/x86_64-linux-gnu
    /usr/include

And this is the list for Clang, using `echo | clang -xc++ -E -v -`:

    /usr/bin/../lib/gcc/x86_64-linux-gnu/13/../../../../include/c++/13
    /usr/bin/../lib/gcc/x86_64-linux-gnu/13/../../../../include/x86_64-linux-gnu/c++/13
    /usr/bin/../lib/gcc/x86_64-linux-gnu/13/../../../../include/c++/13/backward
    /usr/lib/llvm-18/lib/clang/18/include
    /usr/local/include
    /usr/include/x86_64-linux-gnu
    /usr/include

We see that the first and the last three include directories are the
same, but that the middle one—the compiler's private headers include
directory—is different.

So, when we insert libclang's include dirs, we must try to put them in
the same position in CMAKE_CXX_IMPLICIT_INCLUDE_DIRECTORIES, replacing
the one from GCC (or Clang, for completeness).