Add docstring of overloads only once
If an overloaded function or method only has one docstring, or all
overloads (that have a docstring) have the same one, then do not repeat
the docstring, but put it on the "implementation" of the overload.

Mypy doesn't like us adding an "implementation" of overloaded functions
to a stub file [1], OTOH, Ruff doesn't like us adding docstrings to
overloads [2] ...  I believe that adding implementation is more in line
with the reality: at runtime, the Lass-exported function or method only
has 1 docstring and we cannot get access to the "overloaded" functions:
unlike this suggestion [3], `typing.get_overloads(f)` will return an
empty list on overloaded Lass-exported functions.

[1] https://github.com/python/mypy/issues/19128
[2] https://docs.astral.sh/ruff/rules/overload-with-docstring/
[3] https://github.com/astral-sh/ruff/issues/10071#issuecomment-1958557459